Tracker application can now extract some additional information from the message received from the tracker device.
This additional information can be saved to the DB in the new "tracker_history_additional" table.

The XML message received from the Tracker application can contain an "Additional" node as a child of the "History" node.

The additional information may include the LAC/CID, a last known GPS location from the HTTP link embedded within the message, and a message.